+++ src/components/common/Footer/style.tsx
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+export const Root = styled.footer`
+  display: grid;
+  place-items: center;
+  margin-top: auto;
+  padding: 50px 0;
+  font-size: 15px;
+  text-align: center;
+  line-height: 1.5;
+
+  @media (max-width: 768px) {
+    font-size: 13px;
+  }
+`;
