+++ src/components/common/Header/style.tsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const Root = styled.header`
+  position: relative;
+  max-width: 100vw;
+`;
